Refactor modis fixtures to use pytest conftest.py
