+++ .github/workflows/ci.yaml
@@ -16,7 +16,7 @@ jobs:
       - name: Install dependencies
         run: |
           python -m pip install --upgrade pip
-          pip install flake8 flake8-docstrings flake8-debugger flake8-bugbear pytest pytest-lazy-fixture
+          pip install flake8 flake8-docstrings flake8-debugger flake8-bugbear pytest
       - name: Install Satpy
         run: |
           pip install -e .
