+++ continuous_integration/environment.yaml
@@ -40,6 +40,7 @@ dependencies:
   - geoviews
   - pytest
   - pytest-cov
+  - pytest-lazy-fixture
   - fsspec
   - pylibtiff
   - python-geotiepoints
