+++ satpy/tests/reader_tests/conftest.py
@@ -1,0 +1,29 @@
+#!/usr/bin/env python
+# -*- coding: utf-8 -*-
+# Copyright (c) 2021 Satpy developers
+#
+# This file is part of satpy.
+#
+# satpy is free software: you can redistribute it and/or modify it under the
+# terms of the GNU General Public License as published by the Free Software
+# Foundation, either version 3 of the License, or (at your option) any later
+# version.
+#
+# satpy is distributed in the hope that it will be useful, but WITHOUT ANY
+# WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS FOR
+# A PARTICULAR PURPOSE.  See the GNU General Public License for more details.
+#
+# You should have received a copy of the GNU General Public License along with
+# satpy.  If not, see <http://www.gnu.org/licenses/>.
+"""Setup and configuration for all reader tests."""
+
+from ._modis_fixtures import (
+    modis_l1b_nasa_mod021km_file,
+    modis_l1b_imapp_1000m_file,
+    modis_l1b_nasa_mod02hkm_file,
+    modis_l1b_nasa_mod02qkm_file,
+    modis_l1b_nasa_mod03_file,
+    modis_l1b_nasa_1km_mod03_files,
+    modis_l2_nasa_mod35_file,
+    modis_l2_nasa_mod35_mod03_files,
+)
