+++ setup.cfg
@@ -18,6 +18,8 @@ exclude =
     satpy/readers/scatsat1_l2b.py
     satpy/version.py
     satpy/tests/features
+per-file-ignores =
+    satpy/tests/*/conftest.py:F401
 
 [coverage:run]
 relative_files = True
